requirements.txt
# MyFalconAdvisor Requirements
# Generated from pyproject.toml dependencies

# Core AI and LangChain
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.40
openai>=1.12.0

# Financial Data APIs
yfinance>=0.2.18
alpha-vantage>=2.3.1
fredapi>=0.5.0
pandas-datareader>=0.10.0
alpaca-py>=0.21.0

# Data Science and Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Financial Analysis Libraries
pyfolio>=0.9.2
empyrical>=0.5.5
arch>=6.2.0

# Portfolio Optimization
cvxpy>=1.4.0
PyPortfolioOpt>=1.5.5

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# Data Validation and Models
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Async and HTTP
aiohttp>=3.8.0
httpx>=0.24.0

# Configuration and Environment
python-dotenv>=1.0.0

# Background Jobs and Scheduling
schedule>=1.2.0

# CLI Tools
click>=8.1.0
rich>=13.0.0
typer>=0.9.0

# Web API Framework
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.18

# Development and Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
